third_party/blink/web_tests/fast/forms/state-restore-to-non-autocomplete-form-expected.txt

Test to NOT restore form state to a form with autocomplete=off.

Controls in the first form should have their default values:
PASS document.getElementById("input1").value is ""
PASS document.getElementById("textarea1").value is ""
PASS document.getElementById("select1").value is "Mac"
Controls in the second form should have edited values:
PASS document.getElementById("input2").value is "value2"
PASS document.getElementById("textarea2").value is "good"
PASS document.getElementById("select2").value is "BSD"
Controls in the third form should have their default values:
PASS document.getElementById("input3").value is ""
PASS document.getElementById("textarea3").value is ""
PASS document.getElementById("select3").value is "Mac"
PASS successfullyParsed is true

TEST COMPLETE

  
Mac
Windows
BSD
  
Mac
Windows
BSD
  
Mac
Windows
BSD
